# #12133 Speed up twisted.web's HTTP/1.1 server by about 10%