--- a/src/twisted/web/newsfragments/12133.feature
+++ b/src/twisted/web/newsfragments/12133.feature
@@ -1,0 +1,1 @@
+twisted.web's HTTP/1.1 server now runs a little faster, with about 10% lower CPU overhead.