--- a/src/twisted/web/server.py
+++ b/src/twisted/web/server.py
@@ -33,7 +33,7 @@ from twisted.internet.error import AlreadyCalled, AlreadyCancelled
 from twisted.logger import Logger
 from twisted.python import components, failure, reflect
 from twisted.python.compat import nativeString, networkString
-from twisted.python.deprecate import deprecatedModuleAttribute
+from twisted.python.deprecate import deprecated
 from twisted.spread.pb import Copyable, ViewPoint
 from twisted.web import http, iweb, resource, util
 from twisted.web.error import UnsupportedMethod
@@ -50,6 +50,24 @@ __all__ = [
     "NOT_DONE_YET",
     "GzipEncoderFactory",
 ]
+
+
+# backwards compatibility
+@deprecated(
+    Version("Twisted", 12, 1, 0),
+    "twisted.web.http.datetimeToString",
+)
+def date_time_string(dt: int) -> bytes:
+    return http.datetimeToString(dt)
+
+
+@deprecated(
+    Version("Twisted", 12, 1, 0),
+    "twisted.web.http.stringToDatetime",
+)
+def string_date_time(dts: bytes) -> int:
+    return http.stringToDatetime(dts)
+
 
 # Support for other methods may be implemented on a per-resource basis.
 supportedMethods = (b"GET", b"HEAD", b"POST")