--- a/src/twisted/web/test/test_web.py
+++ b/src/twisted/web/test/test_web.py
@@ -1783,53 +1783,6 @@ class LogEscapingTests(unittest.TestCase):
         )
 
 
-class ServerAttributesTests(unittest.TestCase):
-    """
-    Tests that deprecated twisted.web.server attributes raise the appropriate
-    deprecation warnings when used.
-    """
-
-    def test_deprecatedAttributeDateTimeString(self):
-        """
-        twisted.web.server.date_time_string should not be used; instead use
-        twisted.web.http.datetimeToString directly
-        """
-        server.date_time_string(10)
-        warnings = self.flushWarnings(
-            offendingFunctions=[self.test_deprecatedAttributeDateTimeString]
-        )
-
-        self.assertEqual(len(warnings), 1)
-        self.assertEqual(warnings[0]["category"], DeprecationWarning)
-        self.assertEqual(
-            warnings[0]["message"],
-            (
-                "twisted.web.server.date_time_string was deprecated in Twisted "
-                "12.1.0; please use twisted.web.http.datetimeToString instead"
-            ),
-        )
-
-    def test_deprecatedAttributeStringDateTime(self):
-        """
-        twisted.web.server.string_date_time should not be used; instead use
-        twisted.web.http.stringToDatetime directly
-        """
-        server.string_date_time(b"Thursday, 29-Sep-16 17:15:29 GMT")
-        warnings = self.flushWarnings(
-            offendingFunctions=[self.test_deprecatedAttributeStringDateTime]
-        )
-
-        self.assertEqual(len(warnings), 1)
-        self.assertEqual(warnings[0]["category"], DeprecationWarning)
-        self.assertEqual(
-            warnings[0]["message"],
-            (
-                "twisted.web.server.string_date_time was deprecated in Twisted "
-                "12.1.0; please use twisted.web.http.stringToDatetime instead"
-            ),
-        )
-
-
 class ExplicitHTTPFactoryReactor(unittest.TestCase):
     """
     L{http.HTTPFactory} accepts explicit reactor selection.