## Scope and purpose

Fixes #12133

Doing deprecations on module-level is expensive, turns out, because it adds lots of attribute-lookup overhead for _everything_, including non-deprecated code. And other speedups as well.

Before:

```
$ hyperfine --warmup 1 "python http_server_benchmark.py"
Benchmark 1: python http_server_benchmark.py
  Time (mean ± σ):      1.220 s ±  0.010 s    [User: 1.190 s, System: 0.030 s]
  Range (min … max):    1.207 s …  1.238 s    10 runs
```

After:

```
$ hyperfine --warmup 1 "python http_server_benchmark.py"
Benchmark 1: python http_server_benchmark.py
  Time (mean ± σ):      1.105 s ±  0.012 s    [User: 1.081 s, System: 0.024 s]
  Range (min … max):    1.096 s …  1.137 s    10 runs
```

## Contributor Checklist:

This process applies to *all* pull requests - no matter how small.
Have a look at [our developer documentation](https://docs.twisted.org/en/latest/core/development/dev-process.html) before submitting your Pull Request.

Below is a non-exhaustive list (as a reminder):

* The title of the PR should describe the changes and starts with the associated issue number, like “#1234 Brief description”.
* A release notes news fragment file was create in src/twisted/newsfragments/ (see: [Release notes fragments docs.](https://docs.twisted.org/en/latest/core/development/dev-process.html#release-notes-management))
* The automated tests were updated.
* Once all checks are green, request a review by leaving a comment that contains exactly the string `please review`.
  Our bot will trigger the review process, by applying the pending review label
  and requesting a review from the Twisted dev team.