@adiroiban 2024-05-02
@glyph @itamarst about performance testing.

I have approved the request for codspeed ... 

--------


I think that one easy way is to write performance tests, is to write them as trial test.

We can have a pluging to trial that will write a report about the test run.
The report can include the time it takes for the test to run with setup + test + teardown, and just the "inner" test.. all in xunit format.

This data can then be sent to Datadog and we can generate graphs and dashboards.
With datadog import tool, we can add extra attributes for each test.
Datadog offered a free account...as long as we put a banner on our website.

I have implemented this for my private project... but stopped using Datadog as it was too expensive.
I am now working to move the database handling to prometheus or influxdb .

--------

The problem that I can see, is having a reliable / reference run environemnt.
From my past experience, the GitHub Action VMs were not alwasy performing the same.

I can volunteer to host a bare-metal server, dedicated to performance testing, and we can guarantee to have the same results.