@itamarst 2024-05-02
I think the point of codspeed.io is that they provide a reliable runtime environment. We'll see how accurate that claim is.